refactor(tauri): 禁用一些依赖feature以减小bundle体积
移除lib.rs并将主逻辑移至main.rs
清理未使用的依赖项并显式配置各依赖的features
